[no-verify] Make reusable `base_expr` abstraction which can properly recurse for different effects.
